Extract Containerize into a helper function
